Implemented all std opcodes.  Set up function table
All of the opcdoes except for the undocumented ones have been implemented.  The function table was set up.
It could be wrong but I'll have to figure that out in testing.  Working on debug mode now.
That should be accessible from the commandline to enable it which will drop it to a prompt or something.